Allow for changes in TextField to toggle Picker
Feedback from users of a tool I'm developing with material-ui-pickers is they wish they could trigger the date picker without having to click on it. Perhaps onFocus/onChange trigger could be better? I'm willing to discuss this in an issue. I didn't want to open an issue without a solution :)