[typescript] Improve definitions with strictNullChecks disabled (#12895)
While disabling strictNullChecks is still potentially dangerous this removes one false positive that marked the theme as required. Disabling strictNullChecks will still allow the access of theme which will throw at runtime.

The workaround was discovered by @geirsagberg in #12891. I combined this with some necessary documentation so that we can refer to the documentation in future issues because we should not support `strictNullChecks` in the future (DefinitelyTyped isn't either).

Pinging @pelotom in case he has some objections to our tsconfig requirements.
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->