[Table] Fix demo parse rowsPerPage value as an integer (#14848)
The pagination example has a bug where the rowsPerPage field is set to a string when you click on the dropdown. This is because `handleChangeRowsPerPage` method uses `event.target.value` which is a string value.

To replicate this bug. Set page size to `10` then back to `5` and then click the next page button and you'll see that page 2 doesn't show 5 elements instead it shows 10. The `rowsPerPage` value is not recognized and I traced it down to it being set as a string instead of an integer.